requirements.txt
<<<<<<< HEAD
pydantic-settings>=2.3
python-dotenv==1.0.1
# Align fastapi with pyproject and avoid older incompatible pin
fastapi>=0.115,<1.0
uvicorn==0.30.1
pymongo>=4.0
python-multipart>=0.0.7
certifi>=2024.9.0
requests==2.32.3
argon2-cffi>=21.3.0
python-jose[cryptography]>=3.3.0
PyYAML>=6.0
chromadb==0.5.3
streamlit==1.37.1

# LLM & NLP
openai==1.40.0
tiktoken==0.7.0


# Vector DB
faiss-cpu==1.8.0
pinecone-client==3.2.2 # optional, if using Pinecone
weaviate-client==4.6.4 # optional, if using Weaviate


# Data & Utils
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
requests==2.32.3
beautifulsoup4==4.12.3 # for scraping GDPR text
python-docx>=1.1 # provides `docx` module used by ingestion

pdfplumber>=0.11

# protobuf constraint to satisfy google-cloud-aiplatform
# google-cloud-aiplatform requires protobuf>=3.19.5 and <5.0.0dev
protobuf

# Google Cloud Storage client (optional, needed for GCS report uploads)
google-cloud-storage==2.11.0

# Ensure python-docx is at least a modern release (0.8.11 is available on PyPI)
python-docx

# PDF Reports
reportlab==4.2.0


# Testing
pytest==8.2.2
httpx==0.27.0
stripe==5.0.0
# rag==0.1.0
# domain-models==0.1.0
gradio==4.35.0
email_validator
pymongo>=3.6.0
pyflakes

# Optional: sentence-transformers is required for some embedding/semantic search features.
# Note: this may pull in heavy dependencies (torch). If you want a CPU-only install, consider
# pinning a cpu build of torch or installing sentence-transformers inside a separate dev venv.
# dev should turn comment it out for speed build speed
# transformers==4.44.0
# sentence-transformers==3.0.1
# ###


cairosvg
svglib 
lxml
xhtml2pdf
weasyprint
Pango
pydyf
CFFI
tinyhtml5
tinycss2
cssselect2
Pyphen
Pillow
fontTools
=======
adal==1.2.7
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
azure-cognitiveservices-search-imagesearch==2.0.0
azure-common==1.1.28
azure-core==1.30.1
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dnspython==2.7.0
execnb==0.1.5
executing==2.0.1
fastai==2.7.14
fastapi==0.104.1
fastbook==0.0.29
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.19.1
fastprogress==1.0.3
filelock==3.13.4
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
ghapi==1.0.5
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httpx==0.25.1
huggingface-hub==0.22.2
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
ipython-genutils==0.2.0
ipywidgets==7.8.1
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.24
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab-widgets==1.1.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
kiwisolver==1.4.5
langchain==0.2.6
langchain-core==0.2.10
langchain-text-splitters==0.2.2
langcodes==3.3.0
langsmith==0.1.82
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mongomock==4.3.0
motor==3.2.0
mpmath==1.3.0
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.3
nbdev==2.3.13
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.3
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.5
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.5.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.27.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.5.2
pydantic-settings==2.2.1
pydantic_core==2.14.5
pydeck==0.9.1
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.10.1
pyparsing==3.1.2
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.2
sentencepiece==0.2.0
sentinels==1.0.0
sentry-sdk==2.35.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
streamlit==1.36.0
sympy==1.12
tenacity==8.4.2
terminado==0.18.1
thinc==8.2.3
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.24.0.post1
wandb==0.21.1
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==3.6.6
xxhash==3.4.1
yarl==1.9.4
>>>>>>> main
